# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Документация API",
        "title": "Hack",
        "contact": {},
        "version": "1.0"
    },
    "host": "37.18.110.184:3000",
    "basePath": "/",
    "paths": {
...
        "/admin/deleteAdminProfile/id/:userId": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаление аккаунта не админиистратора",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Удаление аккаунта администратора",
                "operationId": "deleteAdminProfile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ИД администратора",
                        "name": "userId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Уровень администратора, где 0-не админ, 1 - админ младщего звена, 3 - старший админ",
                        "name": "adminLevel",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Ид пользователя, которого назначают администратором",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    }
                }
            }
        },
        "/admin/deleteProfile/id/:userId": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаление аккаунта не админиистратора",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Удаление аккаунта",
                "operationId": "deleteProfile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ИД администратора",
                        "name": "userId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Уровень администратора, где 0-не админ, 1 - админ младщего звена, 3 - старший админ",
                        "name": "adminLevel",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Ид пользователя, которого назначают администратором",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    }
                }
            }
        },
        "/admin/promotionAdmin/id/:userId": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Повышение уровня администратора",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Повышение уровня администратора",
                "operationId": "promotionAdmin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ИД администратора",
                        "name": "userId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Уровень администратора, где 0-не админ, 1 - админ младщего звена, 3 - старший админ",
                        "name": "adminLevel",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Ид пользователя, которого назначают администратором",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    }
                }
            }
        },
        "/admin/setAdmin/id/:userId": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Авторизировать пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Назначение администратора",
                "operationId": "setAdmin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ИД администратора",
                        "name": "userId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Уровень администратора, где 0-не админ, 1 - админ младщего звена, 3 - старший админ",
                        "name": "adminLevel",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Ид пользователя, которого назначают администратором",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    }
                }
            }
        },
        "/admin/usSetAdmin/id/:userId": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Снятие с поста администратора",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Снятие с поста администратора",
                "operationId": "unSetAdmin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ИД администратора",
                        "name": "userId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Уровень администратора, где 0-не админ, 1 - админ младщего звена, 3 - старший админ",
                        "name": "adminLevel",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Ид пользователя, которого назначают администратором",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    }
                }
            }
        },
...
        "/getUserInfo": {
            "get": {
                "description": "На вход получаю токен в хедере Authorization",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Проверка токена",
                "operationId": "UserIdentification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal.HackError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "err": {},
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}